requirements.txt
pillow==8.2.0
pygobject==3.40.1
twisted==21.2.0
